+++ src/Support/Http/Middleware/AcceptJson.php
@@ -1,7 +1,5 @@
 <?php
 
-declare(strict_types=1);
-
 namespace Nuwave\Lighthouse\Support\Http\Middleware;
 
 use Closure;