+++ CHANGELOG.md
@@ -15,6 +15,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Include schema directives when running `php artisan lighthouse:validate-schema` https://github.com/nuwave/lighthouse/pull/1494
 - Add ability to query for the existence of relations in where conditions https://github.com/nuwave/lighthouse/pull/1412
 - Handle content types `application/graphql` and `application/x-www-form-urlencoded` properly https://github.com/nuwave/lighthouse/pull/1424
+- Mark directives that can be used more than once per location as `repeatable` https://github.com/nuwave/lighthouse/pull/1529
+- Allow configuring global field middleware directives in `config/lighthouse.php` https://github.com/nuwave/lighthouse/pull/1533
 
 ### Changed
 
@@ -63,6 +65,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Remove the `@middleware` directive, as it violates the boundary between HTTP and GraphQL
   request handling. Use `@guard` or other field middleware directives instead https://github.com/nuwave/lighthouse/pull/1135
 - Remove configuration option `pagination_amount_argument`, it is always `first` now
+
+### Fixed
+
+- Prefix complex conditions with table name to avoid ambiguous SQL https://github.com/nuwave/lighthouse/pull/1530
 
 ## 4.16.1